+++ sphinx_gallery/binder.py
@@ -15,7 +15,7 @@ change in the future.
 
 """
 
-import shutil as sh
+import shutil
 import os
 
 try:
@@ -25,9 +25,13 @@ except NameError:
     unicode = str
 
 from .utils import replace_py_ipynb
+from . import sphinx_compatibility
 
 
-def gen_binder_url(fname, binder_conf):
+logger = sphinx_compatibility.getLogger('sphinx-gallery')
+
+
+def gen_binder_url(fname, binder_conf, gallery_conf):
     """Generate a Binder URL according to the configuration in conf.py.
 
     Parameters
@@ -45,20 +49,36 @@ def gen_binder_url(fname, binder_conf):
     """
     # Build the URL
     fpath_prefix = binder_conf.get('filepath_prefix')
-    binder_fpath = '_downloads/{}'.format(replace_py_ipynb(fname))
+    link_base = binder_conf.get('notebooks_dir')
+
+    # We want to keep the relative path to sub-folders
+    relative_link = os.path.relpath(fname, gallery_conf['src_dir'])
+    path_link = os.path.join(
+        link_base, replace_py_ipynb(relative_link))
+
+    # In case our website is hosted in a sub-folder
     if fpath_prefix is not None:
-        binder_fpath = '/'.join([fpath_prefix.strip('/'), binder_fpath])
+        path_link = '/'.join([fpath_prefix.strip('/'), path_link])
+
+    # Make sure we have the right slashes (in case we're on Windows)
+    path_link = path_link.replace(os.path.sep, '/')
+
+    # Create the URL
     binder_url = binder_conf['url']
     binder_url = '/'.join([binder_conf['url'],
                            'v2', 'gh',
                            binder_conf['org'],
                            binder_conf['repo'],
                            binder_conf['branch']])
-    binder_url += '?filepath={}'.format(binder_fpath)
+
+    if binder_conf.get('use_jupyter_lab', False) is True:
+        binder_url += '?urlpath=lab/tree/{}'.format(path_link)
+    else:
+        binder_url += '?filepath={}'.format(path_link)
     return binder_url
 
 
-def gen_binder_rst(fname, binder_conf):
+def gen_binder_rst(fname, binder_conf, gallery_conf):
     """Generate the RST + link for the Binder badge.
 
     Parameters
@@ -84,7 +104,8 @@ def gen_binder_rst(fname, binder_conf):
     rst : str
         The reStructuredText for the Binder badge that links to this file.
     """
-    binder_url = gen_binder_url(fname, binder_conf)
+    binder_conf = check_binder_conf(binder_conf)
+    binder_url = gen_binder_url(fname, binder_conf, gallery_conf)
 
     rst = (
         "\n"
@@ -95,17 +116,75 @@ def gen_binder_rst(fname, binder_conf):
     return rst
 
 
-def copy_binder_reqs(app):
+def copy_binder_files(app, exception):
+    """Copy all Binder requirements and notebooks files."""
+    if exception is not None:
+        return
+
+    if app.builder.name not in ['html', 'readthedocs']:
+        return
+
+    gallery_conf = app.config.sphinx_gallery_conf
+    binder_conf = check_binder_conf(gallery_conf.get('binder'))
+
+    if not len(binder_conf) > 0:
+        return
+
+    logger.info('copying binder requirements...', color='white')
+    _copy_binder_reqs(app, binder_conf)
+    _copy_binder_notebooks(app)
+
+
+def _copy_binder_reqs(app, binder_conf):
     """Copy Binder requirements files to a "binder" folder in the docs."""
-    binder_conf = app.config.sphinx_gallery_conf['binder']
     path_reqs = binder_conf.get('dependencies')
 
-    binder_folder = os.path.join(app.builder.outdir, 'binder')
+    binder_folder = os.path.join(app.outdir, 'binder')
     if not os.path.isdir(binder_folder):
         os.makedirs(binder_folder)
     for path in path_reqs:
-        sh.copy(os.path.join(app.builder.srcdir, path),
-                binder_folder)
+        shutil.copy(os.path.join(app.srcdir, path), binder_folder)
+
+
+def _remove_ipynb_files(path, contents):
+    """Given a list of files in `contents`, remove all files named `ipynb` or
+    directories named `images` and return the result.
+
+    Used with the `shutil` "ignore" keyword to filter out non-ipynb files."""
+    contents_return = []
+    for entry in contents:
+        if entry.endswith('.ipynb'):
+            # Don't include ipynb files
+            pass
+        elif (entry != "images") and os.path.isdir(os.path.join(path, entry)):
+            # Don't include folders not called "images"
+            pass
+        else:
+            # Keep everything else
+            contents_return.append(entry)
+    return contents_return
+
+
+def _copy_binder_notebooks(app):
+    """Copy Jupyter notebooks to the binder notebooks directory.
+
+    Copy each output gallery directory structure but only including the
+    Jupyter notebook files."""
+
+    gallery_conf = app.config.sphinx_gallery_conf
+    gallery_dirs = gallery_conf.get('gallery_dirs')
+    binder_conf = gallery_conf.get('binder')
+    notebooks_dir = os.path.join(app.outdir, binder_conf.get('notebooks_dir'))
+    shutil.rmtree(notebooks_dir, ignore_errors=True)
+    os.makedirs(notebooks_dir)
+
+    iterator = sphinx_compatibility.status_iterator(
+        gallery_dirs, 'copying binder notebooks...', length=len(gallery_dirs))
+
+    for i_folder in iterator:
+        shutil.copytree(os.path.join(app.srcdir, i_folder),
+                        os.path.join(notebooks_dir, i_folder),
+                        ignore=_remove_ipynb_files)
 
 
 def check_binder_conf(binder_conf):
@@ -119,7 +198,7 @@ def check_binder_conf(binder_conf):
 
     # Ensure all fields are populated
     req_values = ['url', 'org', 'repo', 'branch', 'dependencies']
-    optional_values = ['filepath_prefix']
+    optional_values = ['filepath_prefix', 'notebooks_dir', 'use_jupyter_lab']
     missing_values = []
     for val in req_values:
         if binder_conf.get(val) is None:
@@ -150,6 +229,8 @@ def check_binder_conf(binder_conf):
         raise ValueError("`dependencies` value should be a list of strings. "
                          "Got type {}.".format(type(path_reqs)))
 
+    binder_conf['notebooks_dir'] = binder_conf.get('notebooks_dir',
+                                                   'notebooks')
     path_reqs_filenames = [os.path.basename(ii) for ii in path_reqs]
     if not any(ii in path_reqs_filenames for ii in required_reqs_files):
         raise ValueError(