+++ CHANGES.rst
@@ -7,9 +7,10 @@ git master
 New features
 ''''''''''''
 
-* Added experimental support to auto-generate Binder links for examples via ``binder``
-  config. Note that this API may change in the future. `#244
-  <https://github.com/sphinx-gallery/sphinx-gallery/pull/244>`_
+* Added experimental support to auto-generate Binder links for examples via
+  ``binder`` config. Note that this API may change in the future. `#244
+  <https://github.com/sphinx-gallery/sphinx-gallery/pull/244>`_ and `#371
+  <https://github.com/sphinx-gallery/sphinx-gallery/pull/371>`_.
 * **CSS** Download and binder buttons are now on the right side for large
   screens. This might change downstream CSS layouts.
   "div.sphx-glr-footer-example" can be used to capture these specific